Move `Story Details Presentation` layer to production
