+++ HackrNews/Story Details Presentation/en.lproj/StoryDetails.strings
