+++ HackrNewsTests/Story Details Presentation/StoryDetailsPresentationTests.swift
@@ -5,66 +5,6 @@
 import HackrNews
 import XCTest
 
-struct StoryDetailViewModel {
-    let title: String
-    let author: String
-    let score: String
-    let comments: String
-    let createdAt: String
-    let displayURL: String
-}
-
-class StoryDetailsPresenter {
-    static var title: String {
-        NSLocalizedString(
-            "story_details_title",
-            tableName: "StoryDetails",
-            bundle: Bundle(for: StoryDetailsPresenter.self),
-            value: "",
-            comment: "Story details view title"
-        )
-    }
-
-    private static var score: String {
-        NSLocalizedString(
-            "story_details_score",
-            tableName: "StoryDetails",
-            bundle: Bundle(for: StoryDetailsPresenter.self),
-            value: "",
-            comment: "StoryDetails score text label"
-        )
-    }
-
-    private static var comments: String {
-        NSLocalizedString(
-            "story_details_comments",
-            tableName: "StoryDetails",
-            bundle: Bundle(for: StoryDetailsPresenter.self),
-            value: "",
-            comment: "StoryDetails comments text label"
-        )
-    }
-
-    static func map(
-        _ story: StoryDetail,
-        calendar: Calendar = Calendar(identifier: .gregorian),
-        locale: Locale = .current
-    ) -> StoryDetailViewModel {
-        let formatter = RelativeDateTimeFormatter()
-        formatter.calendar = calendar
-        formatter.locale = locale
-
-        return StoryDetailViewModel(
-            title: story.title,
-            author: story.author,
-            score: String(format: score, story.score),
-            comments: String(format: comments, story.totalComments),
-            createdAt: formatter.localizedString(for: story.createdAt, relativeTo: Date()),
-            displayURL: story.url.host ?? ""
-        )
-    }
-}
-
 final class StoryDetailsPresentationTests: XCTestCase {
     func test_storyDetailsTitle_isLocalized() {
         XCTAssertEqual(StoryDetailsPresenter.title, localized("story_details_title"))
